test(interop): promote verifyInPythonWithFields + parseFieldType to base
These two helpers were copied verbatim across the interop suites; if the
bridge command or fields_hex schema changed, the copies would drift and a
suite would silently stop reflecting reality.

Move both to LXMFInteropTestBase (sibling to the existing verifyInPython),
including the timing log so it's a superset of every copy, and delete the
duplicates from ReactionField/Telemetry/MessageHash/Attachment/Image/Custom
interop tests. Single definition now; net -47 lines. Prunes the imports
left unused by the removals.
